+++ cache-action-entrypoint/webpack.config.d/config.js
@@ -26,9 +26,7 @@ const TerserPlugin = require('terser-webpack-plugin');
 config.optimization = {
   minimizer: [
     new TerserPlugin({
-                       cache: true,
                        parallel: true,
-                       sourceMap: true, // Must be set to true if using source-maps in production
                        terserOptions: {
                          // https://github.com/webpack-contrib/terser-webpack-plugin#terseroptions
                          mangle: false,