+++ wrappers/actions-cache/src/main/kotlin/actions/cache/types.kt
@@ -20,3 +20,4 @@ typealias ValidationError = Error
 
 typealias ReserveCacheError = Error
 
+typealias HttpClientError = Error